requirements.txt
ddgs
langchain_community
langchain_core
markdownify
requests
smolagents
gradio
rank_bm25
llama-cpp-python
fastapi